chore(lib/tutil): move tutil and fix lints
