+++ e2e/cmd/cmd.go
@@ -7,6 +7,7 @@ import (
 	"github.com/omni-network/omni/e2e/app/agent"
 	"github.com/omni-network/omni/e2e/types"
 	libcmd "github.com/omni-network/omni/lib/cmd"
+	"github.com/omni-network/omni/lib/errors"
 	"github.com/omni-network/omni/lib/log"
 
 	cmtdocker "github.com/cometbft/cometbft/test/e2e/pkg/infra/docker"
@@ -88,7 +89,12 @@ func newLogsCmd(def *app.Definition) *cobra.Command {
 		Use:   "logs",
 		Short: "Prints the infrastructure logs (of a previously preserved network)",
 		RunE: func(cmd *cobra.Command, _ []string) error {
-			return cmtdocker.ExecComposeVerbose(cmd.Context(), def.Testnet.Dir, "logs")
+			err := cmtdocker.ExecComposeVerbose(cmd.Context(), def.Testnet.Dir, "logs")
+			if err != nil {
+				return errors.Wrap(err, "executing docker-compose logs")
+			}
+
+			return nil
 		},
 	}
 }
