+++ explorer/indexer/app/config_test.go
@@ -5,9 +5,9 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/omni-network/omni/e2e/tutil"
 	"github.com/omni-network/omni/explorer/indexer/app"
 	"github.com/omni-network/omni/lib/log"
+	"github.com/omni-network/omni/lib/tutil"
 
 	"github.com/stretchr/testify/require"
 )
