+++ halo/genutil/evm/evm_test.go
@@ -3,8 +3,8 @@ package evm_test
 import (
 	"testing"
 
-	"github.com/omni-network/omni/e2e/tutil"
 	"github.com/omni-network/omni/halo/genutil/evm"
+	"github.com/omni-network/omni/lib/tutil"
 
 	_ "github.com/omni-network/omni/halo/app" // To init SDK config.
 )
