+++ halo/genutil/genutil_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/omni-network/omni/e2e/tutil"
 	"github.com/omni-network/omni/halo/genutil"
+	"github.com/omni-network/omni/lib/tutil"
 
 	k1 "github.com/cometbft/cometbft/crypto/secp256k1"
