+++ lib/tutil/golden.go
