+++ lib/tutil/random.go
