+++ lib/tutil/testdata/priv-validator-key.json
