+++ lib/tutil/testdata/priv-validator-state.json
